Move event generator class into scripts/seeds
Removed static seed data
