+++ scripts/seeds/generator_class.rb
